+++ plugin-dev/Source/Sentry/Private/Apple/AppleSentryScope.h
@@ -24,10 +24,6 @@ public:
 	virtual void RemoveTag(const FString& key) override;
 	virtual void SetTags(const TMap<FString, FString>& tags) override;
 	virtual TMap<FString, FString> GetTags() const override;
-	virtual void SetDist(const FString& dist) override;
-	virtual FString GetDist() const override;
-	virtual void SetEnvironment(const FString& environment) override;
-	virtual FString GetEnvironment() const override;
 	virtual void SetFingerprint(const TArray<FString>& fingerprint) override;
 	virtual TArray<FString> GetFingerprint() const override;
 	virtual void SetLevel(ESentryLevel level) override;