+++ plugin-dev/Source/Sentry/Private/Null/NullSentryScope.h
@@ -18,10 +18,6 @@ public:
 	virtual void RemoveTag(const FString& key) override {}
 	virtual void SetTags(const TMap<FString, FString>& tags) override {}
 	virtual TMap<FString, FString> GetTags() const override { return {}; }
-	virtual void SetDist(const FString& dist) override {}
-	virtual FString GetDist() const override { return TEXT(""); }
-	virtual void SetEnvironment(const FString& environment) override {}
-	virtual FString GetEnvironment() const override { return TEXT(""); }
 	virtual void SetFingerprint(const TArray<FString>& fingerprint) override {}
 	virtual TArray<FString> GetFingerprint() const override { return {}; }
 	virtual void SetLevel(ESentryLevel level) override {}