+++ plugin-dev/Source/Sentry/Public/SentryScope.h
@@ -60,22 +60,6 @@ public:
 	UFUNCTION(BlueprintPure, Category = "Sentry")
 	TMap<FString, FString> GetTags() const;
 
-	/** Sets dist in the scope. */
-	UFUNCTION(BlueprintCallable, Category = "Sentry")
-	void SetDist(const FString& Dist);
-
-	/** Gets dist in the scope. */
-	UFUNCTION(BlueprintPure, Category = "Sentry")
-	FString GetDist() const;
-
-	/** Sets environment in the scope. */
-	UFUNCTION(BlueprintCallable, Category = "Sentry")
-	void SetEnvironment(const FString& Environment);
-
-	/** Gets environment in the scope. */
-	UFUNCTION(BlueprintCallable, Category = "Sentry")
-	FString GetEnvironment() const;
-
 	/** Sets fingerprint in the scope. */
 	UFUNCTION(BlueprintCallable, Category = "Sentry")
 	void SetFingerprint(const TArray<FString>& Fingerprint);