+++ plugin-dev/Source/Sentry/Public/SentrySettings.h
@@ -201,6 +201,10 @@ class SENTRY_API USentrySettings : public UObject
 		Meta = (DisplayName = "Environment", ToolTip = "Environment which will be used for enriching events."))
 	FString Environment;
 
+	UPROPERTY(Config, EditAnywhere, BlueprintReadWrite, Category = "General",
+		Meta = (DisplayName = "Distribution", ToolTip = "Distribution which will be used for enriching events."))
+	FString Dist;
+
 	UPROPERTY(Config, EditAnywhere, Category = "General",
 		Meta = (DisplayName = "Sample rate", ToolTip = "Configures the sample rate for error events in the range of 0.0 to 1.0. The default is 1.0 which means that 100% of error events are sent. If set to 0.1 only 10% of error events will be sent. Events are picked randomly.", ClampMin = 0.0f, ClampMax = 1.0f))
 	float SampleRate;